Move pill logic to alchemy feature
