+++ src/features/alchemy/mutators.js
@@ -1,6 +1,7 @@
 import { alchemyState } from './state.js';
 import { ALCHEMY_RECIPES } from './data/recipes.js';
 import { getMaxSlots, getSuccessChance } from './selectors.js';
+import { qCap, clamp } from '../progression/selectors.js';
 
 function slice(state){
   return state.alchemy || alchemyState;
@@ -37,4 +38,25 @@ export function completeBrew(state, index){
 export function unlockRecipe(state, key){
   const alch = slice(state);
   if(!alch.knownRecipes.includes(key)) alch.knownRecipes.push(key);
+}
+
+export function usePill(root, type) {
+  root.pills ??= { qi: 0, body: 0, ward: 0 };
+  if ((root.pills[type] ?? 0) <= 0) return { ok: false, reason: 'none' };
+
+  if (type === 'qi') {
+    const add = Math.floor(qCap(root) * 0.25);
+    root.qi = clamp(root.qi + add, 0, qCap(root));
+  }
+  if (type === 'body') {
+    root.tempAtk = (root.tempAtk || 0) + 4;
+    root.tempArmor = (root.tempArmor || 0) + 3;
+    // NOTE: timer remains in UI for now; long-term move to a timed effect system
+  }
+  if (type === 'ward') {
+    // consumed during breakthrough
+  }
+
+  root.pills[type]--;
+  return { ok: true, type };
 }
