+++ src/features/inventory/mutators.js
@@ -2,7 +2,7 @@ import { S, save } from '../../shared/state.js';
 import { WEAPONS } from '../weaponGeneration/data/weapons.js';
 import { emit } from '../../shared/events.js';
 import { recomputePlayerTotals, canEquip } from './logic.js';
-import { qCap, clamp } from '../progression/selectors.js';
+export { usePill } from '../alchemy/mutators.js'; // deprecated shim
 
 export function addToInventory(item, state = S) {
   state.inventory = state.inventory || [];
@@ -79,23 +79,3 @@ export function unequip(slot, state = S) {
   return payload;
 }
 
-export function usePill(root, type) {
-  root.pills ??= { qi: 0, body: 0, ward: 0 };
-  if ((root.pills[type] ?? 0) <= 0) return { ok: false, reason: 'none' };
-
-  if (type === 'qi') {
-    const add = Math.floor(qCap(root) * 0.25);
-    root.qi = clamp(root.qi + add, 0, qCap(root));
-  }
-  if (type === 'body') {
-    root.tempAtk = (root.tempAtk || 0) + 4;
-    root.tempArmor = (root.tempArmor || 0) + 3;
-    // NOTE: timer remains in UI for now; long-term move to a timed effect system
-  }
-  if (type === 'ward') {
-    // consumed during breakthrough
-  }
-
-  root.pills[type]--;
-  return { ok: true, type };
-}
